@Sword-Smith 2022-07-04
On my `11th Gen Intel(R) Core(TM) i9-11950H @ 2.60GHz`, running Ubuntu 22.04. `rustc 1.61.0 (fe5b13d68 2022-05-18)`

**RP_256**
`main`:
```
time:   [20.789 us 20.801 us 20.815 us]                                   
```

This branch:
```
[15.247 us 15.258 us 15.271 us]
```

**FFT**
`f64_quad/fft_evaluate_poly/with_offset/1048576`
main:
```
time:   [149.50 ms 150.07 ms 150.77 ms]
```

This branch:
```
time:   [127.76 ms 128.98 ms 129.51 ms]
```

All FFT benchmarks are unchanged or faster with this PR.